feat: Add toggle whether to include instructions in prompt
* README.md: Document `mevedel-include-full-instructions`.

* mevedel-instructions.el (mevedel-include-full-instructions)
(mevedel--context, mevedel--directive-llm-prompt):
`mevedel-include-full-instructions` allows the user to control if the
content of directives and references is included in the prompt. If set
to nil, only a reference to the file and line number is included and the
LLM is responsible to run tools to read and process these. This makes
the initial prompt smaller (less tokens used) but, depending on the
model, less reliable in terms of which information is used.