feat: Add ediff session for patch review
Add the option to launch an `ediff` session on the current patch. This
allows the user to review and make any modifications to the patch before
applying it.

* mevedel-instructions.el (mevedel-directive-succeeded-actions-map): Add
`mevedel--ov-actions-ediff`.

(mevedel--ov-actions-dispatch): Add `ediff` action.

* mevedel-utilities.el (mevedel--old-patch-buffer-name)
(mevedel--new-patch-buffer-name, mevedel--ediff-custom-diff-buffer)
(mevedel--original-patch-string, mevedel--ediff-in-progress-p)
(mevedel--ediff-saved-wconf, mevedel--cleanup-ediff-session)
(mevedel--setup-ediff-session, mevedel-ediff-patch)
(mevedel--create-patch-from-ediff, mevedel--store-old-ediff-patch)
(mevedel--create-ediff-custom-patch): Add functionality to run an
`ediff` session on the current patch. Ensures, that the modifications to
the patch are properly tracked and the `ediff` session is cleaned up
afterwards.