feat(mevedel): Add file creation/removal to diff apply buffer
Rename `mevedel--diff-apply-buffer-with-ov-adjustment` to
`mevedel-diff-apply-buffer` and make it an interactive function. It is
now able to handle file creation/deletion along the already implemented
overlay handling.

* README.md: Mention `mevedel-diff-apply-buffer`.

* mevedel-instructions.el (mevedel--directive-next-history)
(mevedel--ov-actions-accept, mevedel--ov-actions-undo):
`mevedel--diff-apply-buffer-with-ov-adjustment` ->
`mevedel-diff-apply-buffer`.

* mevedel-utilities.el (mevedel--diff-apply-buffer-with-ov-adjustment):
`mevedel--diff-apply-buffer-with-ov-adjustment` ->
`mevedel-diff-apply-buffer`.

(mevedel--path-has-suffix-p, mevedel--diff-find-file-operations): Add helper
functions which extract the source and target file relative to the
current workspace and decide if a file needs to be created/deleted.

* mevedel.el (mevedel-process-directives)
(mevedel--auto-apply-patches-maybe, mevedel-uninstall):
`mevedel--diff-apply-buffer-with-ov-adjustment` ->
`mevedel-diff-apply-buffer`.

* tests/test-mevedel-utilities.el: Adjust tests for
`mevedel-diff-apply-buffer`, we need to mock the current workspace.