feat: Add references expansion in chat buffers
The user can now use `@ref` to add a reference to a chat buffer. A
`gptel` transformer will lookup and expand that reference when sending
the prompt to the LLM.

Two types of syntax are supported:
- `@ref:N`, where N is the ID of the reference
- `@ref{TAG QUERY}`, where the references are looked up based on the
  tags

A capf as well as highlighting are provided as well.

* mevedel-instructions.el (mevedel--parse-ref-mentions)
(mevedel--resolve-ref-by-id, mevedel--resolve-refs-by-tag-query)
(mevedel--format-single-reference, mevedel--format-ref-section)
(mevedel--expand-ref-mentions-in-string, mevedel--transform-expand-refs)
(mevedel-ref-capf, mevedel--fontify-ref-id-keyword)
(mevedel--fontify-ref-tag-keyword, mevedel--prettify-ref-mentions):
Added functionality for reference expansion along with
`completion-at-point` function and syntax highlighting.

* mevedel.el (mevedel-install, mevedel-uninstall): Add reference
expansion to install and uninstall commands.